requirements.txt
streamlit>=1.29.0
openai>=1.40.0
pinecone>=5.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=1.1.0
tiktoken>=0.5.2
pyreadline3>=3.5.0

